PartsBin: redesign HBM behaviour
HBM is designed to switch off if screen goes off.
If is activated by a sysfs node, but if it is deactivated,
f.ex. by screen off - the value in sysfs node persists.
It can be only re-enabled, if the sysfs node is re-written -
even, if this is the same value, that currently.
Also we have the issue, that format in sysfs-node can differ.

To get arround this all, we just read the SETTINGS-value - not
the content of sysfs node. Also on screen off KeyHandler switches
the value in SETTINGS to 0.

Change-Id: I4ec3b313901b6d38b8cd8a56d34c2c52f89a910a