PartsBin: update the README
Add package name strings and the string for the Soundtuner activity in the overlay /config.xml for PartsBin.

Those are the default values, which are present in packages_res_deviceparts too, but if the DAX implementation is changed to other files, these strings need to be adapted to make the Soundtuner menu in PartsBin become visible.

First strings make the on/off switch for the Dolby service visible.
Second and third strings make the option to get to the configuration UI for Dolby appear.

Defining none of the strings with a non-OOS package or service, will result in PartsBin NOT displaying the Sound menu at all, as the installed packages are checked by PartsBin based on the strings mentioned in the README and (when undefined) against the ones in res/deviceparts (stock OnePlus package and intent names).

Change-Id: Ica290834df8ba4d559ffa2384b1ce689eebcc227